pkgconfig {zlibbioc}	R Documentation

Retrieve configuration values for PKG_CFLAGS and PKG_LIBS.

Description

Use pkgconfig to retrieve PKG_CFLAGS and PKG_LIBS configuration values to be used in Makevars files. See browseVignettes(package="zlibbioc").

Usage

pkgconfig(opt=c("PKG_CFLAGS", "PKG_LIBS_static", "PKG_LIBS_shared"))

Arguments

opt
A character(1), selected from the values in the default parameter values.

Value

NULL; a side effect is that the configuration string is printed to stdout

Author(s)

Martin Morgan <mtmorgan@fhcrc.org>.

Examples

## Not run: browseVignettes(package="zlibbioc")

pkgconfig("PKG_CFLAGS")
pkgconfig("PKG_LIBS_static")

[Package zlibbioc version 1.28.0 Index]