zlibbioc-package {zlibbioc}	R Documentation

An R packaged zlib-1.2.5

Description

This package is interesting to package developers only, and uses the source code of zlib-1.2.5 to create libraries for systems that do not have these available via other means (most Linux and Mac users should have system-level access to zlib, and no direct need for this package). See the vignette for instructions on use.

Details

Instructions for use are in the vignette, available with browseVignettes(package="zlibbioc")

Author(s)

Author: Martin Morgan

Maintainer: Biocore Team c/o BioC user list <bioconductor@stat.math.ethz.ch>

References

http://zlib.net/

Examples

packageDescription('zlibbioc')

[Package zlibbioc version 1.28.0 Index]